requirements.txt
pandas
yfinance
loguru
pyyaml